Agregando seeders de alumnos, usuarios, notas, corrigiendo migraciones y agregando marca al footer

+++ eva-peron-server/database/migrations/20240508131214-notas.js
@@ -11,7 +11,7 @@ module.exports = {
                 allowNull: false,
             },
             nota: {
-                type: DataTypes.TINYINT(10).UNSIGNED,
+                type: DataTypes.FLOAT.UNSIGNED,
                 allowNull: false,
             },
             alumno_id: {
